# data-collector/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  handler.CollectStockBasicRequest:
    properties:
      mode:
        description: '采集模式: "all", "incremental", "symbol"'
        type: string
      since:
        description: 起始时间（当mode为incremental时使用，格式：2006-01-02）
        type: string
      symbol:
        description: 股票代码（当mode为symbol时使用）
        type: string
    type: object
...
  handler.PolicyListResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      policies:
        items:
          $ref: '#/definitions/models.Policy'
        type: array
      total:
        type: integer
    type: object
  handler.Response:
    properties:
      code:
        description: 响应码
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息
        type: string
      message:
        description: 响应消息
        type: string
    type: object
...
  models.Policy:
    properties:
      content:
        description: 政策内容
        type: string
      created_at:
        description: 创建时间
        type: string
      effective_time:
        description: 生效时间（可选）
        type: string
      id:
        type: string
      impact_level:
        description: 影响级别：high|medium|low
        type: string
      keywords:
        description: 关键词
        items:
          type: string
        type: array
      policy_type:
        description: 政策类型：货币政策|监管政策|交易规则
        type: string
      publish_time:
        description: 发布时间
        type: string
      source:
        description: 发布机构
        type: string
      title:
        description: 政策标题
        type: string
      updated_at:
        description: 更新时间
        type: string
      url:
        description: 原文链接
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 数据采集服务API文档，提供股票、财务、新闻和政策数据采集功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Data Collector API
  version: "1.0"
paths:
...
  /api/v1/policies:
    get:
      consumes:
      - application/json
      description: 根据条件获取政策列表，支持分页
      parameters:
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      - description: 发布机构
        in: query
        name: source
        type: string
      - description: 政策类型
        in: query
        name: policy_type
        type: string
      - description: 影响级别
        in: query
        name: impact_level
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PolicyListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取政策列表
      tags:
      - 政策管理
  /api/v1/policies/{id}:
    get:
      consumes:
      - application/json
      description: 根据政策ID获取政策详细信息
      parameters:
      - description: 政策ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Policy'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据ID获取政策详情
      tags:
      - 政策管理
  /api/v1/policies/search:
    get:
      consumes:
      - application/json
      description: 根据关键词搜索政策
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        required: true
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PolicyListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 搜索政策
      tags:
      - 政策管理
  /api/v1/policies/time-range:
    get:
      consumes:
      - application/json
      description: 根据发布时间范围获取政策列表
      parameters:
      - description: 开始时间
        format: date-time
        in: query
        name: start_time
        required: true
        type: string
      - description: 结束时间
        format: date-time
        in: query
        name: end_time
        required: true
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PolicyListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据时间范围获取政策
      tags:
      - 政策管理
  /api/v1/policies/type/{policy_type}:
    get:
      consumes:
      - application/json
      description: 根据政策类型获取政策列表
      parameters:
      - description: 政策类型
        in: path
        name: policy_type
        required: true
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PolicyListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据政策类型获取政策
      tags:
      - 政策管理
...
  /collect/stock/basic:
    post:
      consumes:
      - application/json
      description: 根据不同模式采集股票基础信息，支持全量采集、增量采集和单个股票采集
      parameters:
      - description: 采集请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CollectStockBasicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 采集成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: 采集股票基础信息
      tags:
      - 股票数据采集
...
schemes:
- http
- https
swagger: "2.0"